% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{geocode_bing}
\alias{geocode_bing}
\title{Geocode addresses and return coordinates in NAD83 / Washington North (EPSG:2285)}
\usage{
geocode_bing(address_line, locality, postal_code, api_key)
}
\arguments{
\item{address_line}{Character vector of street addresses}

\item{locality}{Character vector of cities/localities}

\item{postal_code}{Character vector of postal/ZIP codes}

\item{api_key}{Character string with Bing Maps API key}
}
\value{
A vector with concatenated x and y coordinates (EPSG:2285)
}
\description{
Geocode addresses and return coordinates in NAD83 / Washington North (EPSG:2285)
}
