% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_recognition.R
\name{loc_recognize}
\alias{loc_recognize}
\title{Recognize location type using Bing Maps Location Recognition API}
\usage{
loc_recognize(lng, lat, key_b)
}
\arguments{
\item{lng}{Numeric vector of longitudes}

\item{lat}{Numeric vector of latitudes}

\item{key_b}{Character string with Bing Maps API key}
}
\value{
A character vector with location types ("Residential" or business types),
        or NA for failed requests
}
\description{
Recognize location type using Bing Maps Location Recognition API
}
